"Person" directory renamed to "Application" directory
"Person" directory is currently named "Person" following AB3's
original purpose as an address book application.

InternApply is an application that handles the tracking of
internship "Application", not "Person".

Let's rename the directory to better reflect the purpose of the
"Person" class in InternApply.

A simple refactoring will suffice.

InternApply is still able to succesfully run all previously
implemented tests (Test passed:248 of 248 tests)
