+++ src/main/java/seedu/address/model/AddressBook.java
@@ -5,8 +5,8 @@ import static java.util.Objects.requireNonNull;
 import java.util.List;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.UniquePersonList;
+import seedu.address.model.application.Person;
+import seedu.address.model.application.UniquePersonList;
 
 /**
  * Wraps all data at the address-book level
